==> code/01_export_train_test_GEE.ipynb <==
import ee
import geemap.foliumap as geemap
import json
import os

import pandas as pd
from sklearn import ensemble
import numpy as np

import time

from math import sqrt

from functions import geojson2FeatureCollection, exporto_modis_train_test

ee.Authenticate()
ee.Initialize()

drive.mount('/content/drive', force_remount=True)

train_set = geojson2FeatureCollection('../data/train_data_final.geojson')
test_set = geojson2FeatureCollection('../data/test_data_final.geojson')

print(test_set.getInfo())

modis = ee.ImageCollection("MODIS/MOD09GA_006_NDVI")
region =  ee.FeatureCollection('projects/ee-ldb-deforestacion-montes/assets/Chaco_Seco')

landcover = ee.Image("projects/ee-germanrosati/assets/traj_7years_clust_recode")

for i in [0, 3, 4, 7, 10, 13, 16, 18]:
    
    exporto_modis_train_test(region,
                tipo_ndvi = "anual",
                rango_anios = range(i, i+3, 1),
                file_sufix = str(i)+"-"+str(i+2))
