==> code/02_export_train_test_landcover.ipynb <==
import ee
import json
import os

import geopandas as gpd

import pandas as pd
import numpy as np

import time

from math import sqrt

from functions import geojson2FeatureCollection

ee.Authenticate()
ee.Initialize()

test =  geojson2FeatureCollection("./data/test_data_final.geojson")

landcover = ee.Image("projects/ee-germanrosati/assets/traj_7years_clust_recode")

testing = landcover.sampleRegions(**{
              'collection': test,
              'properties': ['label'],
              'scale': 50,
              'geometries':True
            })

task_test = ee.batch.Export.table.toDrive(**{
          'collection': training,
          'description': 'landcover_test_set',
          'folder':'fundar_deforestacion_input',
          'fileFormat': 'GeoJSON' })
          
task_test.start()
