==> code/04_LR_pipeline.ipynb <==
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV, KFold
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import classification_report

from joblib import dump, load
import geopandas as gpd

import numpy as np

train = gpd.read_file( "./data/train_data_final.geojson")
test =  gpd.read_file("./data/test_data_final.geojson")

X = train.loc[:,'NDVI_2000':'NDVI_2019']
y = train['label']
X_test = test.loc[:,'NDVI_2000':'NDVI_2019']

X

pipe = Pipeline([('preproc_scaling', StandardScaler()),
                 ('lr', LogisticRegression())])

parameters = {'lr__C':np.arange(0.1, 1.1, 0.05),
             'lr__penalty': ['l1', 'l2', 'elasticnet']}

pipe = GridSearchCV(pipe, 
                    parameters,
                    verbose=1,
                    cv=KFold(n_splits=5, shuffle=True, random_state=123),
                    n_jobs=16,
                    #return_train_score=True,
                    scoring='f1_macro'                   
                    #refit=False
                   )

pipe.fit(X, y)

dump(pipe, './models/cv_lr.joblib') 

pipe.best_estimator_

pipe.best_params_

pipe.best_score_

pipe.score(X_test, test['label'])

print(classification_report(test['label'],pipe.best_estimator_.predict(X_test)))


test.to_file("data/predictions_reglog.geojson")
