==> code/05_model_predicts.ipynb <==
from sklearn.metrics import classification_report
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV, KFold
from joblib import dump, load
import geopandas as gpd
import xgboost as xgb
import numpy as np
import pandas as pd

pipe_rf = load('./models/cv_rf.joblib') 
pipe_lr = load('./models/cv_lr.joblib') 
pipe_xgb = load('./models/cv_xgb.joblib') 

train = gpd.read_file( "./data/train_data_final.geojson")
test =  gpd.read_file("./data/test_data_final.geojson")

test.columns

X = train.loc[:,'NDVI_2000':'NDVI_2019']
y = train['label']
X_test = test.loc[:,'NDVI_2000':'NDVI_2019']

best_model_rf = pipe_rf.best_estimator_
best_model_lr = pipe_lr.best_estimator_
best_model_xgb = pipe_xgb.best_estimator_

print("Classification report de RF")
print(classification_report(test['label'],best_model_rf.predict(X_test)))
print("\n")
print("Classification report de reg logística")
print(classification_report(test['label'],best_model_lr.predict(X_test)))
print("\n")
print("Classification report de XGBoost")
#print(classification_report(test['label.x'],best_model_xgb.predict(X_test)))
#print("\n")

test['predictions_rf'] = best_model_rf.predict(X_test)
test['predictions_lr'] = best_model_lr.predict(X_test)
#test['predictions_xgb'] = best_model_xgb.predict(X_test)

test.to_file("data/test_predictions.geojson", driver='GeoJSON')
#Para GEE en .shp
#test.to_file("data/preds/test_predictions.shp")


def exporto_clasif(modelo, nombre):
    report_data = classification_report(test['label'],modelo.predict(X_test), output_dict=True)
    dataframe = pd.DataFrame.from_dict(report_data).reset_index()
    dataframe = dataframe.rename(columns={"index": "metric"})
    
    dataframe.to_csv('data/preds/classification_report_'+nombre+'.csv', index = False)

exporto_clasif(best_model_rf, 'random_forest')
exporto_clasif(best_model_lr, 'reg_log')
#exporto_clasif(best_model_xgb, 'xgb')
