==> code/05_models_selection.ipynb <==
import warnings
warnings.filterwarnings('ignore')

from functions import define_final_model
from joblib import load

pipe_rf = load('./models/cv_rf.joblib') 
pipe_lr = load('./models/cv_lr.joblib') 
pipe_xgb = load('./models/cv_xgb_2.joblib') 

define_final_model(pipe_rf, 'random_forest')
define_final_model(pipe_lr, 'logistic_reg')
define_final_model(pipe_xgb, 'xgb')
