==> code/06_model_predicts.ipynb <==
import warnings
warnings.filterwarnings('ignore')

from sklearn.metrics import classification_report

from joblib import load

import geopandas as gpd

import xgboost as xgb

import numpy as np
import pandas as pd

rf = load('./models/random_forest.joblib') 
lr = load('./models/logistic_reg.joblib') 
pipe_xgb = load('./models/xgb.joblib') 

train = gpd.read_file( "./data/train_data_final.geojson")
test =  gpd.read_file("./data/test_data_final.geojson")

years_cols = ['year_'+str(i) for i in range(0,20)]

cols_to_update_train = years_cols + train.columns[20:].tolist()
train.columns = cols_to_update_train

cols_to_update_test = years_cols + test.columns[20:].tolist()
test.columns = cols_to_update_test

X = train.loc[:,'year_0':'year_19']
y = train['label']
X_test = test.loc[:,'year_0':'year_19']

print("Classification report de RF")
print(classification_report(test['label'],rf.predict(X_test)))
print("\n")
print("Classification report de reg logística")
print(classification_report(test['label'],lr.predict(X_test)))
print("\n")
print("Classification report de XGBoost")
print(classification_report(test['label'],pipe_xgb.predict(X_test)))
print("\n")

test['predictions_rf'] = rf.predict(X_test)
test['predictions_lr'] = lr.predict(X_test)
test['predictions_xgb'] = pipe_xgb.predict(X_test)

def exporto_clasif(modelo, nombre):
    
    report_data = classification_report(test['label'],modelo.predict(X_test), output_dict=True)
    
    dataframe = pd.DataFrame.from_dict(report_data).reset_index()
    dataframe = dataframe.rename(columns={"index": "metric"})
    
    dataframe.to_csv('data/preds/classification_report_'+nombre+'.csv', index = False)

exporto_clasif(rf, 'random_forest')
exporto_clasif(lr, 'reg_log')
exporto_clasif(pipe_xgb, 'xgb')

test.to_file("data/test_predictions.geojson", driver='GeoJSON')
