==> code/07_model_errors.ipynb <==
import geopandas as gpd
import numpy as np
import pygeos

gpd.options.use_pygeos=True

predictions = gpd.read_file("data/landcover_test_set.geojson")

predictions

errores_RF = predictions[predictions['label'] != predictions['predictions_rf']]

errores_RF['b1'].value_counts(normalize=True)
