==> code/10_test_region.ipynb <==
import ee
from functions import geojson2FeatureCollection, exporto_modis, proc_data, run_predictions
import glob
import pandas as pd
from functools import reduce
import geopandas as gpd

ee.Authenticate()
ee.Initialize()

region = geojson2FeatureCollection("../data/test_region/points_entrerios.geojson")

for i in [0, 3, 4, 7, 10, 13, 16, 18]:
    
    exporto_modis(region,
                rango_anios = range(i, i+3, 1),
                file_sufix = str(i)+"-"+str(i+2))

proc_data(path = '../data/test_region/ndvi/*.geojson',
          path_output = '../data/test_region/data_final.geojson')

predicted_data = run_predictions('../data/test_region/data_final.geojson',
               model = "both")

predicted_data['RF_pred'].value_counts()

predicted_data.to_file("../data/test_region/entre_rios_data_final_preds.geojson")
